Code was refactoring and now this project has a log file. Continue to resolve errors.txt
