+++ velhia.game/errors.txt
@@ -1,5 +1,8 @@
-Não possue um arquivo de Log
+Ok - Não possue um arquivo de Log
+Ok - Não está deletando tudo no final do Play()
+Ok - Não sei se está aguardando o retorno da API
+Ok - Completar o arquivo de log
+
 Não está validando se toda sociedade & ambiente estão batendo os dados
-Não sei se está aguardando o retorno da API
-Não está deletando tudo no final do Play()
+
 Não dando baixa na partida quando checa se o agente morreu
