+++ velhia.game/errors/invalid_match.py
@@ -1,0 +1,3 @@
+class InvalidMatch(Exception):
+    """ Raised when there is an incongruity in the society"""
+    pass
