+++ velhia.game/handlers/statistical_algorithm_handler.py
@@ -1,0 +1,37 @@
+import sys
+import json
+from datetime import datetime
+from classes.statistical_algorithm import StatisticalAlgorithm
+
+
+def get_lastest_sa(db):
+    """
+    Get the lastest statistical algorithm obj in the database
+    or create if it not exists
+
+    Usage
+    >>> sa = get_lastest_sa()
+    >>> print(sa)
+
+    <classes.statistical_algorithm.StatisticalAlgorithm object at 0x7fe6de3287d0>
+    <classes.statistical_algorithm.StatisticalAlgorithm object at 0x7fe6de3287d0>
+    """
+
+    try:
+        if len(db.get_last(1).json()) == 0:
+            sa = StatisticalAlgorithm(db.create(json.dumps({
+                "birth": datetime.now().ctime(),
+                "memory": [],
+                "matchs": 0,
+                "victories": 0,
+                "defeats": 0,
+                "draw": 0
+            })).json(), ('X', 1), ('O', 0))
+        else:
+            sa = StatisticalAlgorithm(
+                db.get_last(1).json()[0],
+                ('X', 1), ('O', 0))
+
+        return sa
+    except:
+        raise SystemError
